Update strawberry authentication
- introspect token once per request
- split authentication/authorization classes
- rename graphql_authentication.py to strawberry.py
