+++ tests/strawberry/test_authenticated_field.py
